Rename check_version11 function to check_nocd
The check_version11 function checks a byte in Freelancer.exe to determine whether the 1.1 patch is installed. However, this check is invalid since the 1.1 patch doesn't modify Freelancer.exe. Additionally, all functionalities in flplusplus work just fine without the 1.1 patch installed (from my testing at least). Though one issue that could occur with the plugin is if the No-CD EXE is NOT used, because the offsets will be different. So the check has been renamed to No-CD. Also FALSE will now be returned when the No-CD EXE is not detected to ensure the DLL gets unloaded (saves memory).