Get base addresses for soundmanager and soundstreamer
Using the hardcoded virtual addresses will work most of the time. However, I've seen cases where one of the DLLs get moved (loaded at a different base address). As a result, the patch will be applied at the incorrect address, and in some cases the game may even crash. Therefore, it's better to play it safe and always get the base address to guarantee the patch will be correct.